+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## dev
 - Show diff in trails
+- Simplify structure for attributes in audit entry view
 
 ## 1.2.5
 - Add user ip callback